Adjust jpl_alloc to use a signed argument
It's literally impossible to allocate more than a 63-bit number of
bytes, so I think it's safe to make this signed and therefore more
uniform with JPL.

+++ runtime.h
@@ -14,7 +14,7 @@ struct args {
 int main(int, char**);
 
 void fail_assertion(char*);
-void *jpl_alloc(size_t);
+void *jpl_alloc(int64_t);
 double get_time(void);
 
 void show(char*, void*);
